requirements.txt
aws-cdk-lib==2.43.1
constructs==10.1.112
aws-cdk.aws-lambda-python-alpha>=2.42.1a0
boto3==1.24.80
black==22.8.0
requests==2.28.1
notion-client==1.0
python-telegram-bot==13.14